requirements.txt
selenium
langchain
langchain-core
langchain-openai
python-dotenv